finalização dos testes unitários da camada de serviços

+++ src/main/java/com/graphql/project/dtos/CreateOrderProduct.java
@@ -9,5 +9,5 @@ import java.math.BigDecimal;
 public record CreateOrderProduct(int orderId,
                                  int productCode,
                                  int qty,
-                                 BigDecimal priceEach) {
+                                 int priceEach) {
 }
