+++ src/main/java/com/graphql/project/dtos/UpdateOrderProduct.java
@@ -2,5 +2,5 @@ package com.graphql.project.dtos;
 
 import java.math.BigDecimal;
 
-public record UpdateOrderProduct(int qnty, BigDecimal priceEach) {
+public record UpdateOrderProduct(int qnty, int priceEach) {
 }
